requirements.txt
# Core LangChain packages
langchain==0.1.20
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-community==0.0.38

# LangGraph
langgraph==0.0.55

# Playwright
playwright==1.44.0

# PDF Processing
PyPDF2==3.0.1

# Data Models
pydantic==2.7.1
pydantic-core==2.18.2

# Environment Variables
python-dotenv==1.0.1

# OpenAI
openai==1.30.1

# Web Scraping (for locator inspector)
beautifulsoup4==4.12.3
requests==2.32.3

# Testing
pytest==8.2.1
pytest-playwright==0.4.4